# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Exocortex Web API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task",
        "operationId": "get_task_api_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/asset-init": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Submit Asset Init",
        "operationId": "submit_asset_init_api_tasks_asset_init_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_asset_init_api_tasks_asset_init_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummaryModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/compress-preview": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Submit Compress Preview",
        "operationId": "submit_compress_preview_api_tasks_compress_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompressTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummaryModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/compress-execute": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Submit Compress Execute",
        "operationId": "submit_compress_execute_api_tasks_compress_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompressTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummaryModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/group-dive": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Submit Group Dive",
        "operationId": "submit_group_dive_api_tasks_group_dive_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummaryModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/ask-tutor": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Submit Ask Tutor",
        "operationId": "submit_ask_tutor_api_tasks_ask_tutor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskTutorWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummaryModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/re-tutor": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Submit Re Tutor",
        "operationId": "submit_re_tutor_api_tasks_re_tutor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReTutorWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummaryModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/integrate": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Submit Integrate",
        "operationId": "submit_integrate_api_tasks_integrate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummaryModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/bug-finder": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Submit Bug Finder",
        "operationId": "submit_bug_finder_api_tasks_bug_finder_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_bug_finder_api_tasks_bug_finder_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummaryModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/student-note": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Submit Student Note",
        "operationId": "submit_student_note_api_tasks_student_note_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummaryModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/fix-latex": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Submit Fix Latex",
        "operationId": "submit_fix_latex_api_tasks_fix_latex_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FixLatexWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummaryModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AskTutorWorkflowRequest": {
        "properties": {
          "assetName": {
            "type": "string",
            "title": "Assetname"
          },
          "groupIdx": {
            "type": "integer",
            "title": "Groupidx"
          },
          "tutorIdx": {
            "type": "integer",
            "title": "Tutoridx"
          },
          "question": {
            "type": "string",
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "assetName",
          "groupIdx",
          "tutorIdx",
          "question"
        ],
        "title": "AskTutorWorkflowRequest"
      },
...
      "Body_submit_asset_init_api_tasks_asset_init_post": {
        "properties": {
          "source_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Source File"
          },
          "markdown_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Markdown File"
          },
          "content_list_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Content List File"
          },
          "asset_name": {
            "type": "string",
            "title": "Asset Name"
          },
          "asset_subfolder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Subfolder"
          }
        },
        "type": "object",
        "required": [
          "source_file",
          "markdown_file",
          "content_list_file",
          "asset_name"
        ],
        "title": "Body_submit_asset_init_api_tasks_asset_init_post"
      },
      "Body_submit_bug_finder_api_tasks_bug_finder_post": {
        "properties": {
          "assetName": {
            "type": "string",
            "title": "Assetname"
          },
          "groupIdx": {
            "type": "integer",
            "title": "Groupidx"
          },
          "tutorIdx": {
            "type": "integer",
            "title": "Tutoridx"
          },
          "manuscript_files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Manuscript Files"
          }
        },
        "type": "object",
        "required": [
          "assetName",
          "groupIdx",
          "tutorIdx",
          "manuscript_files"
        ],
        "title": "Body_submit_bug_finder_api_tasks_bug_finder_post"
      },
      "CompressTaskRequest": {
        "properties": {
          "assetName": {
            "type": "string",
            "title": "Assetname"
          },
          "fractionRect": {
            "$ref": "#/components/schemas/RectModel"
          },
          "ratio": {
            "type": "integer",
            "title": "Ratio"
          },
          "compressScale": {
            "type": "number",
            "title": "Compressscale",
            "default": 1.0
          },
          "drawBadge": {
            "type": "boolean",
            "title": "Drawbadge",
            "default": true
          },
          "badgePosition": {
            "type": "string",
            "title": "Badgeposition",
            "default": "top_left"
          }
        },
        "type": "object",
        "required": [
          "assetName",
          "fractionRect",
          "ratio"
        ],
        "title": "CompressTaskRequest"
      },
...
      "FixLatexWorkflowRequest": {
        "properties": {
          "assetName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assetname"
          },
          "markdownPath": {
            "type": "string",
            "title": "Markdownpath"
          }
        },
        "type": "object",
        "required": [
          "markdownPath"
        ],
        "title": "FixLatexWorkflowRequest"
      },
...
      "GroupWorkflowRequest": {
        "properties": {
          "assetName": {
            "type": "string",
            "title": "Assetname"
          },
          "groupIdx": {
            "type": "integer",
            "title": "Groupidx"
          }
        },
        "type": "object",
        "required": [
          "assetName",
          "groupIdx"
        ],
        "title": "GroupWorkflowRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReTutorWorkflowRequest": {
        "properties": {
          "assetName": {
            "type": "string",
            "title": "Assetname"
          },
          "groupIdx": {
            "type": "integer",
            "title": "Groupidx"
          },
          "tutorIdx": {
            "type": "integer",
            "title": "Tutoridx"
          },
          "question": {
            "type": "string",
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "assetName",
          "groupIdx",
          "tutorIdx",
          "question"
        ],
        "title": "ReTutorWorkflowRequest"
      },
      "RectModel": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "width": {
            "type": "number",
            "title": "Width"
          },
          "height": {
            "type": "number",
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "RectModel"
      },
...
      "TaskDetailModel": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "assetName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assetname"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/TaskEventModel"
            },
            "type": "array",
            "title": "Events"
          },
          "latestEvent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskEventModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskResultModel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "title",
          "createdAt",
          "updatedAt"
        ],
        "title": "TaskDetailModel"
      },
      "TaskEventModel": {
        "properties": {
          "taskId": {
            "type": "string",
            "title": "Taskid"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "assetName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assetname"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "eventType": {
            "type": "string",
            "title": "Eventtype"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "artifactPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifactpath"
          },
          "payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskFailurePayloadModel"
              },
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "boolean"
                    },
                    {
                      "items": {},
                      "type": "array"
                    },
                    {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "taskId",
          "kind",
          "status",
          "eventType",
          "message",
          "timestamp"
        ],
        "title": "TaskEventModel"
      },
...
      "TaskResultModel": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "artifactPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifactpath"
          },
          "payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskFailurePayloadModel"
              },
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "boolean"
                    },
                    {
                      "items": {},
                      "type": "array"
                    },
                    {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "TaskResultModel"
      },
      "TaskSummaryModel": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "assetName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assetname"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "title",
          "createdAt",
          "updatedAt"
        ],
        "title": "TaskSummaryModel"
      },
...
      "TutorWorkflowRequest": {
        "properties": {
          "assetName": {
            "type": "string",
            "title": "Assetname"
          },
          "groupIdx": {
            "type": "integer",
            "title": "Groupidx"
          },
          "tutorIdx": {
            "type": "integer",
            "title": "Tutoridx"
          }
        },
        "type": "object",
        "required": [
          "assetName",
          "groupIdx",
          "tutorIdx"
        ],
        "title": "TutorWorkflowRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}